rename EAForumWrapped2023Page.tsx
