+++ packages/lesswrong/lib/components.ts
@@ -49,7 +49,7 @@ importComponent("EACollectionCard", () => require('../components/ea-forum/EAColl
 importComponent("EAHomeHandbook", () => require('../components/ea-forum/EAHomeHandbook'));
 importComponent("EAHomeRightHandSide", () => require('../components/ea-forum/EAHomeRightHandSide'));
 importComponent("EAForumWrapped2022Page", () => require('../components/ea-forum/wrapped/EAForumWrapped2022Page'));
-importComponent("EAForumWrapped2023Page", () => require('../components/ea-forum/wrapped/EAForumWrapped2023Page'));
+importComponent("EAForumWrapped2023Page", () => require('../components/ea-forum/wrapped/EAForumWrappedPage'));
 importComponent("EASurveyBanner", () => require('../components/ea-forum/EASurveyBanner'));
 importComponent("Digests", () => require('../components/ea-forum/digest/Digests'));
 importComponent("EditDigest", () => require('../components/ea-forum/digest/EditDigest'));
